Move ProblematicClassGenerator out
